♻️ refactor: mysql, redis와 Repository 계층 분리
