+++ feed-crawler/src/repository/rss.repository.ts
@@ -1,0 +1,12 @@
+import { mysqlConnection } from "src/common/mysql-access";
+import { RssObj } from "../common/types";
+
+class RssRepository {
+  public async selectAllRss(): Promise<RssObj[]> {
+    const query = `SELECT id, rss_url as rssUrl, name as blogName, blog_platform as blogPlatform
+        FROM rss_accept`;
+    return mysqlConnection.executeQuery(query);
+  }
+}
+
+export const rssRepository = new RssRepository();
